@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
-apply plugin: 'android-apt'
+apply plugin: 'com.jakewharton.butterknife'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "com.justwayward.reader"
         minSdkVersion 14
-        targetSdkVersion 22
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -47,37 +46,50 @@ android {
             jniLibs.srcDirs = ['libs']
         }
     }
+    compileOptions {
+        sourceCompatibility '1.8'
+        targetCompatibility '1.8'
+    }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile project(':EasyAdapterLibrary')
-    compile 'com.android.support:appcompat-v7:23.4.0'
-    compile 'com.android.support:design:23.4.0'
-    compile 'com.android.support:cardview-v7:23.4.0'
-    // Net
-    compile 'com.squareup.retrofit2:retrofit:2.1.0'
-    compile 'com.squareup.retrofit2:adapter-rxjava:2.0.1'
-    compile 'com.squareup.retrofit2:converter-gson:2.0.0-beta4'
-    compile 'com.squareup.retrofit2:converter-scalars:2.0.0-beta4'
-    compile 'com.squareup.okhttp3:okhttp:3.2.0'
-    // Annotation
-    compile 'com.google.dagger:dagger:2.0.2'
-    apt 'com.google.dagger:dagger-compiler:2.0.2'
-    provided 'org.glassfish:javax.annotation:10.0-b28'
-    compile 'com.jakewharton:butterknife:7.0.1'
-    // Rx
-    compile 'io.reactivex:rxandroid:1.1.0'
-    compile 'io.reactivex:rxjava:1.1.5'
-    // Bus
-    compile 'org.greenrobot:eventbus:3.0.0'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+//    implementation project(path: ':EasyAdapterLibrary', configuration: 'default')
+    implementation project( path: ':EasyAdapterLibrary' )
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
 
-    compile 'com.commit451:PhotoView:1.2.4'
-    compile 'org.ccil.cowan.tagsoup:tagsoup:1.2.1'
+    // Net
+    implementation 'com.squareup.retrofit2:retrofit:2.1.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava:2.0.1'
+    implementation 'com.squareup.retrofit2:converter-gson:2.0.0-beta4'
+    implementation 'com.squareup.retrofit2:converter-scalars:2.0.0-beta4'
+    implementation 'com.squareup.okhttp3:okhttp:3.2.0'
+    // Annotation
+//    implementation 'com.google.dagger:dagger:2.+'
+//    annotationProcessor 'com.google.dagger:dagger-compiler:2.+'
+//    implementation 'com.google.dagger:dagger-android:2.+' // If you're using classes in dagger.android
+//    implementation 'com.google.dagger:dagger-android-support:2.+' // if you use the support libraries
+//    annotationProcessor 'com.google.dagger:dagger-android-processor:2.+'
+
+    implementation 'com.google.dagger:dagger:2.+'
+    annotationProcessor 'com.google.dagger:dagger-compiler:2.+'
+    compileOnly 'org.glassfish:javax.annotation:10.0-b28'
+    implementation 'com.jakewharton:butterknife:10.2.3'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.3'
+    // Rx
+    implementation 'io.reactivex:rxandroid:1.1.0'
+    implementation 'io.reactivex:rxjava:1.1.5'
+    // Bus
+    implementation 'org.greenrobot:eventbus:3.0.0'
+
+    implementation 'com.commit451:PhotoView:1.2.4'
+    implementation 'org.ccil.cowan.tagsoup:tagsoup:1.2.1'
 
     // leakcanary
-    debugCompile 'com.squareup.leakcanary:leakcanary-android:1.5'
-    releaseCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.5'
-    testCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.5'
+//    debugImplementation  'com.squareup.leakcanary:leakcanary-android:1.5'
+//    releaseImplementation  'com.squareup.leakcanary:leakcanary-android-no-op:1.5'
+//    testImplementation 'com.squareup.leakcanary:leakcanary-android-no-op:1.5'
 
 }
