@@ -16,8 +16,8 @@
 package com.justwayward.reader.base;
 
 import android.content.Context;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.LinearLayoutManager;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import android.view.View;
 
 import com.justwayward.reader.R;
@@ -30,7 +30,7 @@ import java.lang.reflect.Constructor;
 
 import javax.inject.Inject;
 
-import butterknife.Bind;
+import butterknife.BindView;
 
 /**
  * @author lfh.
@@ -41,7 +41,7 @@ public abstract class BaseRVFragment<T1 extends BaseContract.BasePresenter, T2> 
     @Inject
     protected T1 mPresenter;
 
-    @Bind(R.id.recyclerview)
+    @BindView(R.id.recyclerview)
     protected EasyRecyclerView mRecyclerView;
     protected RecyclerArrayAdapter<T2> mAdapter;
 
