@@ -18,10 +18,10 @@ package com.justwayward.reader.ui.activity;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AlertDialog;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.app.AlertDialog;
 import android.view.Menu;
 import android.view.MenuItem;
 
@@ -40,7 +40,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import butterknife.Bind;
+import butterknife.BindView;
 
 /**
  * 书籍详情 社区
@@ -62,9 +62,9 @@ public class BookDetailCommunityActivity extends BaseActivity {
     private String title;
     private int index;
 
-    @Bind(R.id.indicatorSubRank)
+    @BindView(R.id.indicatorSubRank)
     RVPIndicator mIndicator;
-    @Bind(R.id.viewpagerSubRank)
+    @BindView(R.id.viewpagerSubRank)
     ViewPager mViewPager;
 
     private List<Fragment> mTabContents;
