@@ -17,10 +17,10 @@ package com.justwayward.reader.ui.activity;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.support.v7.widget.ListPopupWindow;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.widget.ListPopupWindow;
 import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -47,7 +47,7 @@ import java.util.List;
 
 import javax.inject.Inject;
 
-import butterknife.Bind;
+import butterknife.BindView;
 
 /**
  * @author yuyh.
@@ -62,9 +62,9 @@ public class SubCategoryListActivity extends BaseActivity implements SubCategory
 
     private String currentMinor = "";
 
-    @Bind(R.id.indicatorSub)
+    @BindView(R.id.indicatorSub)
     RVPIndicator mIndicator;
-    @Bind(R.id.viewpagerSub)
+    @BindView(R.id.viewpagerSub)
     ViewPager mViewPager;
 
     @Inject
