@@ -18,7 +18,7 @@ package com.justwayward.reader.ui.activity;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.text.TextUtils;
 import android.view.View;
 import android.widget.TextView;
@@ -26,21 +26,15 @@ import android.widget.TextView;
 import com.justwayward.reader.R;
 import com.justwayward.reader.api.BookApi;
 import com.justwayward.reader.base.BaseActivity;
-import com.justwayward.reader.bean.BookMixAToc;
-import com.justwayward.reader.bean.ChapterRead;
 import com.justwayward.reader.component.AppComponent;
 import com.justwayward.reader.component.DaggerMainComponent;
-import com.justwayward.reader.ui.contract.BookReadContract;
-import com.justwayward.reader.ui.presenter.BookReadPresenter;
 import com.justwayward.reader.utils.NetworkUtils;
 import com.justwayward.reader.wifitransfer.Defaults;
 import com.justwayward.reader.wifitransfer.ServerRunner;
 
-import java.util.List;
-
 import javax.inject.Inject;
 
-import butterknife.Bind;
+import butterknife.BindView;
 import butterknife.OnClick;
 
 /**
@@ -52,12 +46,12 @@ public class WifiBookActivity extends BaseActivity {
         context.startActivity(new Intent(context, WifiBookActivity.class));
     }
 
-    @Bind(R.id.mTvWifiName)
+    @BindView(R.id.mTvWifiName)
     TextView mTvWifiName;
-    @Bind(R.id.mTvWifiIp)
+    @BindView(R.id.mTvWifiIp)
     TextView mTvWifiIp;
 
-    @Bind(R.id.tvRetry)
+    @BindView(R.id.tvRetry)
     TextView tvRetry;
 
     @Inject
