@@ -3,8 +3,8 @@ package com.justwayward.reader.ui.fragment;
 import android.graphics.PorterDuff;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
+import androidx.fragment.app.Fragment;
+import androidx.core.content.ContextCompat;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -19,7 +19,7 @@ import com.justwayward.reader.ui.activity.ReadEPubActivity;
 import com.justwayward.reader.view.epubview.ObservableWebView;
 import com.justwayward.reader.view.epubview.VerticalSeekbar;
 
-import butterknife.Bind;
+import butterknife.BindView;
 import nl.siegmann.epublib.domain.Book;
 
 /**
@@ -33,9 +33,9 @@ public class EPubReaderFragment extends BaseFragment {
     private static final String BUNDLE_EPUB_FILE_NAME = "filename";
     private static final String BUNDLE_IS_SMIL_AVAILABLE = "smilavailable";
 
-    @Bind(R.id.scrollSeekbar)
+    @BindView(R.id.scrollSeekbar)
     VerticalSeekbar mScrollSeekbar;
-    @Bind(R.id.contentWebView)
+    @BindView(R.id.contentWebView)
     ObservableWebView mWebview;
 
     private int mPosition = -1;
