@@ -16,9 +16,9 @@
 package com.justwayward.reader.ui.fragment;
 
 import android.content.Intent;
-import android.net.Uri;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.justwayward.reader.R;
@@ -35,7 +35,7 @@ import com.justwayward.reader.view.SupportDividerItemDecoration;
 import java.util.ArrayList;
 import java.util.List;
 
-import butterknife.Bind;
+import butterknife.BindView;
 
 /**
  * 发现
@@ -45,7 +45,7 @@ import butterknife.Bind;
  */
 public class FindFragment extends BaseFragment implements OnRvItemClickListener<FindBean> {
 
-    @Bind(R.id.recyclerview)
+    @BindView(R.id.recyclerview)
     RecyclerView mRecyclerView;
 
     private FindAdapter mAdapter;
