@@ -16,8 +16,8 @@
 package com.justwayward.reader.view;
 
 import android.content.Context;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.ListPopupWindow;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.widget.ListPopupWindow;
 import android.util.AttributeSet;
 import android.view.LayoutInflater;
 import android.view.View;
