@@ -18,7 +18,7 @@ package com.justwayward.reader.view.loadding;
 import android.app.Activity;
 import android.app.Dialog;
 import android.content.Context;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.view.View;
 import android.view.ViewGroup;
 
