@@ -16,7 +16,7 @@
 package com.justwayward.reader.view.pdfview;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.view.MotionEvent;
 import android.view.View;
 
